move shortcut dialog to ui module
